clean up Update
